backend/pyproject.toml
[tool.poetry]
package-mode = false
name = "ft-transcendence"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.1"
daphne = "^4.1.2"
djangorestframework = "^3.15.2"
twisted = {extras = ["http2", "tls"], version = "^24.7.0"}
psycopg = {extras = ["binary"], version = "^3.2.2"}
django-health-check = "^3.18.3"
djangorestframework-simplejwt = "^5.3.1"
django-otp = "^1.5.4"
12factor-vault = "^0.1.23"
django-postgresql-setrole = "^1.0.12"
pillow = "^10.4.0"
requests = "^2.32.3"
django-cors-headers = "^4.5.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
